Changed file name to signup-requests and fixed console warning
